서버 에러 사항이 삼켜지는 문제 리펙토링( 기존 timer 삭제 및 룸허브의 각 룸의 함수를 실행시켜주는 로직 안에서 게임 진행 로직 돌아가도록 수정)

+++ HelloServer/FSM/StateMachine.cs
@@ -76,7 +76,7 @@ namespace Jay.FSM
             return CurrentState is T;
         }
 
-        public void Tick(float deltaTime)
+        public void Tick(int deltaTime)
         {
             if (CurrentState is IUpdatableState updatable)
             {
